backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.1
email-validator>=2.2.0
requests>=2.32.0
groq>=0.11.0
llama-index-core>=0.11.0
llama-index-vector-stores-faiss>=0.2.0
llama-index-embeddings-huggingface>=0.3.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
